backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sentence-transformers==2.2.2
faiss-cpu==1.7.4
google-generativeai==0.3.2
numpy==1.24.3
python-dotenv==1.0.0
PyPDF2==3.0.1